Put definitions of symmetric matrices to avoid import loops
